Fix some conversion warnings.
Exclude definitions from my_config.h that conflict with server's config.h
(most of ma_config.h is not used by connector anyway)

Avoid overriding server's SIGN_TARGET() macro, if used as subproject.
Make WITH_UNIT_TESTS exclude unit tests, like it is for the server.

+++ unittest/libmariadb/ps_bugs.c
@@ -2699,7 +2699,7 @@ static int test_bug5194(MYSQL *mysql)
     }
     *query_ptr= '\0';
 
-    rc= mysql_stmt_prepare(stmt, query, query_ptr - query);
+    rc= mysql_stmt_prepare(stmt, query, (ulong)(query_ptr - query));
 
     if (rc && nrows * COLUMN_COUNT > uint16_max) /* expected error */
       break;
